Implement basic type inference of bril programs
To do this, we simply go through each instruction of the program and try to
determine what the type of the arguments should be based on the op we're doing,
as well as the type of the destination. If there is a conflict with existing
inferred types, we raise an Exception.

Note that `id` is tricky because it could be that we have something like
`x = id y` but we don't know the type of `y` until later in the program
since we're going through sequentially. Thus we can just keep doing
type inference until we can't make any more new inferences.