Fix bug with ambiguity in label vs untyped op
Consider the two programs:
```
main {
l:
  x = const 5
}
```
which creates a label `l` and a variable `x` of type `int` with a value of 5 and
```
main {
  l: x = const 5
}
```
which creates a variable `l` of type `x` with a value of 5.
This is ambiguous for the parser. To fix this, I changed the parser
`type` nonterminal to only be "int" or "bool" instead of any CNAME.