[to_ssa] Fix bug in rename()
The use of `defaultdict()` and `stack.update(old_stack)` meant that any
stacks that were empty at the top of `_rename()` weren't cleared at the
bottom. This caused use of undefined variables inserted into `phi`
instead of `__undefined`.

The fix is to copy the stack at the top of _rename(). This is just as
efficient as before (not very efficient).