brilift: Remove target special-casing for ARM
Brilift AOT didn't previously support macOS-on-ARM because
`cranelift-object` didn't, which in turn is because the `object` crate
was missing some relevant features. It seems like that support has been
added, possibly in this PR but I'm not sure:
https://github.com/gimli-rs/object/pull/553

Anyway, stuff seems to be working on ARM now, which is cool. Now we can
remove the hacks to insist on Rosetta2 when on Apple Silicon.