Reshuffle `compile` function
The `CompileArgs` struct wasn't very helpful, so I made these actual
arguments to the function. The `OptLevel` enum was also sort of in the
wrong place: this is about command-line parsing rather than being about
an actual useful interface (since we will pass this as a string to
Cranelift anyway).