Add Blum Blum Shub benchmark
This benchmark prints the least significant bit of each elm of a given
BBS sequence. To spot-check for correctness, the tested program output
sequence is matched against example output in the Wikipedia entry for
the BBS PRNG (when p=11, q=23, seed=3, and sequence_length=6).

No additional testing of this program (e.g. on other BBS sequences) was
performed.